# frontend/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "eWorldCup Manager API",
    "description": "API for managing round-robin tournament schedules",
    "version": "v1"
  },
  "paths": {
...
    "/Tournament/start": {
      "post": {
        "tags": [
          "Tournament"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TournamentStartRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TournamentStartRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TournamentStartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentStartResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Tournament/{tournamentId}/status": {
      "get": {
        "tags": [
          "Tournament"
        ],
        "parameters": [
          {
            "name": "tournamentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Tournament/{tournamentId}/play": {
      "post": {
        "tags": [
          "Tournament"
        ],
        "parameters": [
          {
            "name": "tournamentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayRoundRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayRoundRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PlayRoundRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayRoundResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Tournament/{tournamentId}/advance": {
      "post": {
        "tags": [
          "Tournament"
        ],
        "parameters": [
          {
            "name": "tournamentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvanceRoundResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Tournament/{tournamentId}/final": {
      "get": {
        "tags": [
          "Tournament"
        ],
        "parameters": [
          {
            "name": "tournamentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinalResultResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AdvanceRoundResponse": {
        "required": [
          "completedRound",
          "matchesSimulated",
          "message",
          "tournamentComplete",
          "tournamentId"
        ],
        "type": "object",
        "properties": {
          "tournamentId": {
            "type": "integer",
            "format": "int32"
          },
          "completedRound": {
            "type": "integer",
            "format": "int32"
          },
          "matchesSimulated": {
            "type": "integer",
            "format": "int32"
          },
          "tournamentComplete": {
            "type": "boolean"
          },
          "nextRound": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "nextOpponent": {
            "type": "string",
            "nullable": true
          },
          "scoreboard": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScoreboardEntryDto"
            },
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "FinalResultResponse": {
        "required": [
          "playerName",
          "playerRank",
          "playerWon",
          "totalRounds",
          "tournamentId",
          "winner",
          "winnerIndex",
          "winnerPoints"
        ],
        "type": "object",
        "properties": {
          "tournamentId": {
            "type": "integer",
            "format": "int32"
          },
          "playerName": {
            "type": "string",
            "nullable": true
          },
          "totalRounds": {
            "type": "integer",
            "format": "int32"
          },
          "winner": {
            "type": "string",
            "nullable": true
          },
          "winnerIndex": {
            "type": "integer",
            "format": "int32"
          },
          "winnerPoints": {
            "type": "integer",
            "format": "int32"
          },
          "playerWon": {
            "type": "boolean"
          },
          "playerRank": {
            "type": "integer",
            "format": "int32"
          },
          "finalScoreboard": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScoreboardEntryDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "MatchStatusDto": {
        "required": [
          "gameRoundNumber",
          "matchStatus",
          "opponentIndex",
          "opponentName",
          "opponentWins",
          "playerWins"
        ],
        "type": "object",
        "properties": {
          "opponentName": {
            "type": "string",
            "nullable": true
          },
          "opponentIndex": {
            "type": "integer",
            "format": "int32"
          },
          "gameRoundNumber": {
            "type": "integer",
            "format": "int32"
          },
          "playerWins": {
            "type": "integer",
            "format": "int32"
          },
          "opponentWins": {
            "type": "integer",
            "format": "int32"
          },
          "matchStatus": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PlayRoundRequest": {
        "required": [
          "choice"
        ],
        "type": "object",
        "properties": {
          "choice": {
            "$ref": "#/components/schemas/RpsChoice"
          }
        },
        "additionalProperties": false
      },
      "PlayRoundResponse": {
        "required": [
          "gameRoundNumber",
          "matchComplete",
          "message",
          "opponentChoice",
          "opponentWins",
          "playerChoice",
          "playerWins",
          "result"
        ],
        "type": "object",
        "properties": {
          "gameRoundNumber": {
            "type": "integer",
            "format": "int32"
          },
          "playerChoice": {
            "$ref": "#/components/schemas/RpsChoice"
          },
          "opponentChoice": {
            "$ref": "#/components/schemas/RpsChoice"
          },
          "result": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "playerWins": {
            "type": "integer",
            "format": "int32"
          },
          "opponentWins": {
            "type": "integer",
            "format": "int32"
          },
          "matchComplete": {
            "type": "boolean"
          },
          "matchWinner": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "RpsChoice": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "ScoreboardEntryDto": {
        "required": [
          "losses",
          "playerIndex",
          "playerName",
          "points",
          "wins"
        ],
        "type": "object",
        "properties": {
          "playerIndex": {
            "type": "integer",
            "format": "int32"
          },
          "playerName": {
            "type": "string",
            "nullable": true
          },
          "wins": {
            "type": "integer",
            "format": "int32"
          },
          "losses": {
            "type": "integer",
            "format": "int32"
          },
          "points": {
            "type": "integer",
            "format": "int32"
          },
          "rank": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "TournamentStartRequest": {
        "required": [
          "playerName",
          "totalPlayers"
        ],
        "type": "object",
        "properties": {
          "playerName": {
            "type": "string",
            "nullable": true
          },
          "totalPlayers": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "TournamentStartResponse": {
        "required": [
          "firstOpponent",
          "firstOpponentIndex",
          "maxRounds",
          "playerName",
          "tournamentId"
        ],
        "type": "object",
        "properties": {
          "tournamentId": {
            "type": "integer",
            "format": "int32"
          },
          "playerName": {
            "type": "string",
            "nullable": true
          },
          "totalPlayers": {
            "type": "integer",
            "format": "int32"
          },
          "maxRounds": {
            "type": "integer",
            "format": "int32"
          },
          "firstOpponent": {
            "type": "string",
            "nullable": true
          },
          "firstOpponentIndex": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "TournamentStatusResponse": {
        "required": [
          "currentRound",
          "maxRounds",
          "status",
          "tournamentId"
        ],
        "type": "object",
        "properties": {
          "tournamentId": {
            "type": "integer",
            "format": "int32"
          },
          "currentRound": {
            "type": "integer",
            "format": "int32"
          },
          "maxRounds": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "currentMatch": {
            "$ref": "#/components/schemas/MatchStatusDto"
          },
          "scoreboard": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScoreboardEntryDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}